+++ lib/widgets/settings/habits/habits_type_card.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:lotti/classes/entity_definitions.dart';
 import 'package:lotti/themes/theme.dart';
 import 'package:lotti/widgets/journal/entry_tools.dart';
+import 'package:lotti/widgets/settings/categories/categories_type_card.dart';
 import 'package:lotti/widgets/settings/settings_card.dart';
 import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
 
@@ -9,17 +10,20 @@ class HabitsTypeCard extends StatelessWidget {
   const HabitsTypeCard({
     required this.item,
     required this.index,
+    required this.color,
     super.key,
   });
 
   final HabitDefinition item;
   final int index;
+  final Color color;
 
   @override
   Widget build(BuildContext context) {
     return SettingsNavCard(
       path: '/settings/habits/${item.id}',
       title: item.name,
+      leading: CategoryColorIcon(color),
       trailing: Row(
         mainAxisSize: MainAxisSize.min,
         children: [