+++ test/pages/settings/habits/habits_page_test.dart
@@ -1,9 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:lotti/classes/entity_definitions.dart';
 import 'package:lotti/database/database.dart';
 import 'package:lotti/get_it.dart';
 import 'package:lotti/pages/settings/habits/habits_page.dart';
 import 'package:lotti/themes/themes_service.dart';
+import 'package:mocktail/mocktail.dart';
 
 import '../../../mocks/mocks.dart';
 import '../../../test_data/test_data.dart';
@@ -17,6 +19,12 @@ void main() {
   group('Habits Page Widget Tests - ', () {
     setUp(() {
       mockJournalDb = mockJournalDbWithHabits([habitFlossing]);
+
+      when(mockJournalDb.watchCategories).thenAnswer(
+        (_) => Stream<List<CategoryDefinition>>.fromIterable([
+          [categoryMindfulness]
+        ]),
+      );
 
       getIt
         ..registerSingleton<JournalDb>(mockJournalDb)